aim-guardrails/requirements.txt
flask>=2.3.0
requests>=2.31.0
detoxify>=0.6.0
presidio-analyzer>=2.2.0
presidio-anonymizer>=2.2.0
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
prometheus-client>=0.19.0
torch>=2.0.0
transformers>=4.35.0
pyyaml>=6.0

